+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed errors for models that have missing values and optional property
 - Fixed topological sort for execution planning
 - Fixed processing non-optional object property
+- Fixed bugs in `drepr.utils.attr_data` and `rdfgraph_writer` that write after abort records.
 
 ## [1.1.3] - 2024-03-26